refactor: streamline distribution update tests
